requirements.txt
psycopg2-binary==2.9.6
pyfaktory
python-dotenv~=1.0
requests~=2.32